+++ sport/app/cricket/views/fragments/cricketMatchSummary.scala.html
@@ -1,14 +1,17 @@
 @import java.time.format.DateTimeFormatter
 @import java.time.ZoneId
-@(theMatch: cricketModel.Match, matchUrl: String)(implicit request: RequestHeader)
 @import common.LinkTo
+@import common.Chronos
 @import views.support.GuDateFormatLegacy
+
+
+@(theMatch: cricketModel.Match, matchUrl: String)(implicit request: RequestHeader)
 
 <div class="sport-summary sport-summary--cricket" itemscope itemtype="http://schema.org/SportsEvent">
 
     <h2 class="u-h">
         <time class="u-h" datetime="@theMatch.gameDate.format(DateTimeFormatter.ISO_DATE)" data-timestamp="@theMatch.gameDate.atZone(ZoneId.of("UTC")).toInstant().toEpochMilli()">
-        @GuDateFormatLegacy.apply2(theMatch.gameDate, "d MMM y")
+        @GuDateFormatLegacy(Chronos.toDateTime(theMatch.gameDate), "d MMM y")
         </time>
         @theMatch.competitionName, @theMatch.venueName
     </h2>
